Fix the installation of KyTea's library
When setting up the building environment on Travis CI, install the library
and headers under /usr/local.  Include /usr/local/lib in LD_LIBRARY_PATH
just in case the path is not included in the environment variable.  The previous
approach requires users to set up the environment varibles related to cgo.
That is not easy if they are not familiar with cgo.  This change avoids
setting the variables as long as they have correctly installed KyTea library under
the system in which a compler finds the headers, and a dynamic linker finds
the library.